requirements.txt
flet==0.80.5
flet-geolocator==0.80.5
flet-audio-recorder==0.80.5
supabase==2.11.0
requests==2.32.3
python-dotenv==1.2.1
supabase-auth==2.27.2
postgrest==0.19.3
realtime==2.27.2
httpx==0.28.1
websockets==13.1
packaging==26.0
watchfiles==1.1.1
